Write collections to public directory.
